[Controller] Leader controller should stop push monitor after transiting to standby
This RB fixes a bug that controllers still monitor offline push status after
transiting from leader to standy; as a result, the standby controller still
act on push status changes. For example, when all partitions of a hybrid store
report END_OF_PUSH_RECEIVED, controller would kick off buffer replay; with this
bug, if there is leader handover after a push starts but not ends yet, more
than 1 controller would try to kick off buffer replay, which confuses servers
since servers would receive different StartOfBufferReplay/TopicSwitch messages.

Also, PushMonitor#stopMonitorOfflinePush function had an assumption that
stopping monitoring should also delete push status in ZK. This RB refactors
this function to include an extra parameter to control whether to delete
push status or not.

RB=2047800

RB=2047800
BUG=VENG-5734
G=voldemort-reviewers
R=xnma,goliu,sdwu,zfan,anaberez,xyin,jlliu,fvillega,shuhliu,yurzhou,smaji,nithakka,zpolicze
A=sdwu

+++ venice-common/src/main/java/com/linkedin/venice/pushmonitor/AbstractPushMonitor.java
@@ -155,7 +155,7 @@ public abstract class AbstractPushMonitor
         if (existingStatus.equals(ExecutionStatus.ERROR)) {
           logger.info("The previous push status for topic: " + kafkaTopic + " is 'ERROR',"
               + " and the new push will clean up the previous 'ERROR' push status");
-          cleanupPushStatus(topicToPushMap.get(kafkaTopic));
+          cleanupPushStatus(topicToPushMap.get(kafkaTopic), true);
         } else {
           throw new VeniceException("Push status has already been created for topic:" + kafkaTopic + " in cluster:" + clusterName);
         }
@@ -173,7 +173,7 @@ public abstract class AbstractPushMonitor
   }
 
   @Override
-  public void stopMonitorOfflinePush(String kafkaTopic) {
+  public void stopMonitorOfflinePush(String kafkaTopic, boolean deletePushStatus) {
     logger.info("Stopping monitoring push on topic:" + kafkaTopic);
     pushMonitorWriteLock.lock();
     try {
@@ -188,9 +188,26 @@ public abstract class AbstractPushMonitor
         String storeName = Version.parseStoreFromKafkaTopicName(pushStatus.getKafkaTopic());
         retireOldErrorPushes(storeName);
       } else {
-        cleanupPushStatus(pushStatus);
+        cleanupPushStatus(pushStatus, deletePushStatus);
       }
       logger.info("Stopped monitoring push on topic:" + kafkaTopic);
+    } finally {
+      pushMonitorWriteLock.unlock();
+    }
+  }
+
+  @Override
+  public void stopAllMonitoring() {
+    logger.info("Stopping monitoring push for all topics.");
+    pushMonitorWriteLock.lock();
+    try {
+      for (Map.Entry<String, OfflinePushStatus> entry : topicToPushMap.entrySet()) {
+        String kafkaTopic = entry.getKey();
+        stopMonitorOfflinePush(kafkaTopic, false);
+      }
+      logger.info("Successfully stopped monitoring push for all topics.");
+    } catch (Exception e) {
+      logger.error("Error when stopping monitoring push for all topics", e);
     } finally {
       pushMonitorWriteLock.unlock();
     }
@@ -204,7 +221,7 @@ public abstract class AbstractPushMonitor
           .filter(topic -> Version.parseStoreFromKafkaTopicName(topic).equals(storeName))
           .collect(Collectors.toList());
 
-      topicList.forEach(topic -> cleanupPushStatus(topicToPushMap.get(topic)));
+      topicList.forEach(topic -> cleanupPushStatus(topicToPushMap.get(topic), true));
     } finally {
       pushMonitorWriteLock.unlock();
     }
@@ -292,11 +309,13 @@ public abstract class AbstractPushMonitor
   /**
    * this is to clear legacy push statuses
    */
-  private void cleanupPushStatus(OfflinePushStatus offlinePushStatus) {
+  private void cleanupPushStatus(OfflinePushStatus offlinePushStatus, boolean deletePushStatus) {
     pushMonitorWriteLock.lock();
     try {
       topicToPushMap.remove(offlinePushStatus.getKafkaTopic());
-      offlinePushAccessor.deleteOfflinePushStatusAndItsPartitionStatuses(offlinePushStatus);
+      if (deletePushStatus) {
+        offlinePushAccessor.deleteOfflinePushStatusAndItsPartitionStatuses(offlinePushStatus);
+      }
     } catch (Exception e) {
       logger.warn("Could not delete legacy push status: " + offlinePushStatus.getKafkaTopic(), e);
     } finally {
@@ -333,7 +352,7 @@ public abstract class AbstractPushMonitor
       // Make sure we do boxing; List.remove(primitive int) treats the primitive int as index
       versionNums.remove(Integer.valueOf(errorVersion));
 
-      cleanupPushStatus(errorPushStatus);
+      cleanupPushStatus(errorPushStatus, true);
     }
   }
